==> 03_rothc_FB.RMD <==
---
output: html_document
editor_options: 
  chunk_output_type: console
---

Experiment: C-FB

#------------------------------------------------------------------------------#
# Script to run rothc with data from literature research and database: EJPSoil_BUFACAP_template.xlsx
#
# This script only treats experiment: Farako-Ba, Burkina Faso (SSA)
#
# created by T. Ruhnau (22.10.2024)
#------------------------------------------------------------------------------#

Read packages and input data -> data comes from 02_climate_data.R and 01_auto_database_FB.R
```{r Packages}
## load packages
library(SoilR)
library(readr)
library(ggplot2)

# Read dataframes from both scripts before continue (first parts)
# > 02_climate_data_FB.R holds all important data on climate
# > 01_auto_database_FB.R holds all important data on soils
```
#-- Preperation ---------------------------------------------------------------#

The information of vegetation periods has to be taken directly from the literature or NDVI values, if not available in LTFEs

Split climate data up for historical spin-up (1970-1990) and prediction model (1991-2024)  
(The split-up assignment is random and follows no peer-reviewed basis)

According to Peña Cabriales et al. (2006) on pg. 59 it says: 
"one rainy season per year, starting in May-June and ending in October." 
> Filter is set accordingly.

Always run in the beginning
```{r Climate monthly means}
# Monthly means from climate_data_FB.R
mm1 <- read.csv2("output/farako_ba/climate/mm_1970_1990.csv", header = TRUE) %>% select(-1)
mm2 <- read.csv2("output/farako_ba/climate/mm_1991_2024.csv", header = TRUE) %>% select(-1)

# Filter Month and Experiment for Spinup section (1970 - 1990)
mm1_spinup <- mm1 %>%
  # filter(Month >= 5 & Month <= 10) %>% 
  filter(Experiment == "Farako-Ba") # Choose the experiment

# Filter Month and Experiment for Section1 Rain (1991 - 2024)
mm2_sec1 <- mm2 %>% # This will be used for Model Section
  filter(Month >= 5 & Month <= 10) %>% 
  filter(Experiment == "Farako-Ba") # Choose the experiment

# Filter Month and Experiment for Section2 Dry (1991 - 2024)
mm2_sec2 <- mm2 %>% # This will be used for Model Section
  filter(Month %in% c(11:12, 1:4)) %>% # to filter months through year
  filter(Experiment == "Farako-Ba") # Choose the experiment

rm(mm1, mm2)
```
```{r Read soil data}
# Specify the directory where your CSV files are located
csv_dir <- "output/farako_ba/" # change dir to different LTFEs

# Get a list of all CSV files in the directory
csv_files <- list.files(path = csv_dir, pattern = "*.csv", full.names = TRUE)

# Read all CSV files into a list of dataframes
df_list <- map(csv_files, ~read_csv2(.) %>% select(-1)) # -1 to get rid of first column

# Name the dataframes based on the file names (without extension)
names(df_list) <- tools::file_path_sans_ext(basename(csv_files))

# Unpack list to global environment
list2env(df_list, .GlobalEnv)

rm(df_list)
```

#-- Spin-up section -----------------------------------------------------------#

```{r Climate Spin-up}
## Climate data for Farako-ba, Burkina Faso for 1970-1990 (ERA5)

# Mean monthly temperature for 1970-1990
Temp_Spinup <- mm1_spinup %>% 
  filter(Variable == "Temp") %>% 
  select(Month, MeanValue) %>% 
  rename(Temp = MeanValue)

# Mean monthly precipitation for 1970-1990
Prec_Spinup <- mm1_spinup %>% 
  filter(Variable == "Prec") %>% 
  select(Month, MeanValue) %>% 
  rename(Prec = MeanValue)

# Mean monthly evaporation for 1970-1990
Evap_Spinup <- mm1_spinup %>% 
  filter(Variable == "Evapo") %>% 
  select(Month, MeanValue) %>% 
  rename(Evap = MeanValue)
```

```{r Soil Spin-up}
# Prepare soil data
soil.thick_Spinup=20 # Same as model
SOC_Spinup=17.6        # Change to fit initial SOC stock of model ---> 17.6 for C-FB
clay_Spinup=7        # Same as model
Cinputs_Spinup=2.96     # Change to historical land-use --> 2.96 for C-FB
DPM.RPM = 1.44       # Change if land-use is different to model
FYM = 0              # Input factor for Farmyard manure by time

# Create vector of time steps to run spin-up
years_Spinup=seq(from=0, to=500, by=1/12)
```

```{r Prepare Spin-up}
# Calculate effects of climate on decomposition
fT_Spinup=fT.RothC(Temp_Spinup[,2]) # Temperature effects per month 
fW_Spinup=fW.RothC(P=(Prec_Spinup[,2]), E=(Evap_Spinup[,2]),
            S.Thick = soil.thick_Spinup, pClay = clay_Spinup,
            pE = 1.0, bare = FALSE)$b # Rainy season = vegetation covering the soil (bare = F)

# Replace negative values in fW with 0
fW_Spinup[fW_Spinup < 0] <- 0 # high Evapo values leads to error in model, still not sure if this is the right way to handle it

# >>Add comment<<
xi.frame_Spinup=data.frame(years_Spinup,rep(fT_Spinup*fW_Spinup,length.out=length(years_Spinup)))

# Estimating IOM using Falloon method
FallIOM=0.049*SOC_Spinup^(1.139) 
```

```{r Model Spin-up}
# Run the model Section1
Model_Spinup=RothCModel(t=years_Spinup,C0=c(DPM=0, RPM=0, BIO=0, HUM=0, IOM=FallIOM), 
                  In=Cinputs_Spinup, clay=clay_Spinup, xi=xi.frame_Spinup,
                  DR = DPM.RPM, FYM = FYM)

# Calculates SOC stocks for each pool
Ct_Spinup=getC(Model_Spinup) 

# Final pool size spin-up
poolSize_Spinup=as.numeric(tail(Ct_Spinup,1)) # Uses the last values to show final pool size
names(poolSize_Spinup)<-c("DPM", "RPM", "BIO", "HUM", "IOM") # Is that correct?

# Print pool size for spin-up section
poolSize_Spinup

# Sum the carbon stocks for total spin up
total_C_spinup <- sum(poolSize_Spinup)

# Combine total carbon stocks into a data frame
total_C_df <- data.frame(
    Model = c("Spinup"),
    Total_C = c(sum(poolSize_Spinup))
)

# Print total sum of C after spin-up
print(total_C_df)
```

#-- Prediction section --------------------------------------------------------#

```{r Sec.1 Climate Model}
### Climate data for Farako-ba, Burkina Faso 
## Section1 Rainy Season (May-Oct) for 1991-2024 (ERA5)

# Mean monthly temperature for 1991-2024 (ERA5)
Temp_Model1 <- mm2_sec1 %>% 
  filter(Variable == "Temp") %>% 
  select(Month, MeanValue) %>% 
  rename(Temp = MeanValue)
# Mean monthly precipitation for 1991-2024 (ERA5)
Prec_Model1 <- mm2_sec1 %>% 
  filter(Variable == "Prec") %>% 
  select(Month, MeanValue) %>% 
  rename(Prec = MeanValue)
# Mean monthly evaporation for 1991-2024 (ERA5)
Evap_Model1 <- mm2_sec1 %>% 
  filter(Variable == "Evapo") %>% 
  select(Month, MeanValue) %>% 
  rename(Evap = MeanValue)
```

```{r Sec.2 Climate Model}
### Climate data for Farako-ba, Burkina Faso for 1991-2024 (ERA5)
## Section2 Dry Season (Nov-Apr) for 1991-2024 (ERA5)

# Mean monthly temperature for 1991-2024 (ERA5)
Temp_Model2 <- mm2_sec2 %>% 
  filter(Variable == "Temp") %>% 
  select(Month, MeanValue) %>% 
  rename(Temp = MeanValue)
# Mean monthly precipitation for 1991-2024 (ERA5)
Prec_Model2 <- mm2_sec2 %>% 
  filter(Variable == "Prec") %>% 
  select(Month, MeanValue) %>% 
  rename(Prec = MeanValue)
# Mean monthly evaporation for 1991-2024 (ERA5)
Evap_Model2 <- mm2_sec2 %>% 
  filter(Variable == "Evapo") %>% 
  select(Month, MeanValue) %>% 
  rename(Evap = MeanValue)
```

```{r Soil Prep C-FB}
### This chunk selects all parameters for the treatment C-FB (control site in Farako-Ba)
## For further information see literature in database

# Calculate soil.thick
fb_datasoil <- fb_datasoil %>% 
  mutate(soil.thick = Depth_to_cm - Depth_from_cm)

# Get initial SOC stock 
fb_SOC <- fb_datasoil %>% 
  filter(tID == "C-FB") %>% 
  select(eID, tID, Sample_year, SOC_stock_Mg.ha) # Use first year as initial SOC stock, in this case 1993 = 16.06

# Get clay content
fb_clay <- fb_soiltype %>% 
  filter(Bottom_depth_cm == 20) # Use for clay model

# Get carbon residue input
fb_cinput <- fb_datacrop %>% 
  select(eID, tID, Residues_above_Mg.ha, DC_Comment) %>% 
  filter(tID == "C-FB") %>% 
  mutate(MeanResidues = mean(Residues_above_Mg.ha))# Use average for cinputs_spinup
```

```{r Sec.1 Soil Model}
### Section1: Vegetation (Rainy Season) = TRUE (bare = FALSE)
## soil data
soil.thick_Model1 = fb_datasoil$soil.thick[1]   # Soil thickness (organic layer topsoil), in cm
SOC_Model1 = 16.06    ### Spin-up should fit this initial SOC stock # Soil organic carbon in Mg/ha 
clay_Model1 = 7       # Percent clay
Cinputs_Model1 = fb_cinput$MeanResidues[1]       # Annual C inputs to soil in Mg/ha/yr
DPM.RPM = 1.44         # DPM/RPM --> DR = arable: 1.44, grasslands: 0.67, forest: 0.25 (Gottschalk et al. 2012)
FYM = 0                # FYM allocation: DPM: 49%, RPM: 49%, HUM: 2% (Weihermueller et al. 2013)

# create vector of time steps to run simulation
years_Model1 = seq(1/6,20,by=1/6) # change sequence according to rain season
```

```{r Sec.2 Soil Model}
### Section2: Vegetation (Dry Season) = FALSE (bare = TRUE)
## soil data
# soil.thick_Model1 = fb_datasoil$soil.thick[1]   # Soil thickness (organic layer topsoil), in cm
# SOC_Model1 = 16.06        # Soil organic carbon in Mg/ha 
# clay_Model1 = 7       # Percent clay
Cinputs_Model2 = 0      # Annual C inputs to soil in Mg/ha/yr
# DPM.RPM = 1.44         # DPM/RPM --> DR = arable: 1.44, grasslands: 0.67, forest: 0.25 (Gottschalk et al. 2012)
# FYM = 0                # FYM allocation: DPM: 49%, RPM: 49%, HUM: 2% (Weihermueller et al. 2013)

# create vector of time steps to run simulation
years_Model2 = seq(1/6,20,by=1/6) # change sequence according to vegetation period
```

```{r Prepare Model}
#------------------------------- Section1 -------------------------------------#

# Calculate effects of climate on decomposition
fT_Model1=fT.RothC(Temp_Model1[,2]) # Temperature effects per month 
fW_Model1=fW.RothC(P=(Prec_Model1[,2]), E=(Evap_Model1[,2]),
            S.Thick = soil.thick_Model1, pClay = clay_Model1,
            pE = 1.0, bare = FALSE)$b # Rainy season = vegetation covering the soil (bare = F)

# Replace negative values in fW with 0
fW_Model1[fW_Model1 < 0] <- 0 # high Evapo values leads to error in model

xi.frame_Model1=data.frame(years_Model1,rep(fT_Model1*fW_Model1,length.out=length(years_Model1)))

#------------------------------- Section2 -------------------------------------#

# Calculate effects of climate on decomposition
fT_Model2=fT.RothC(Temp_Model2[,2]) # Temperature effects 
fW_Model2=fW.RothC(P=(Prec_Model2[,2]), E=(Evap_Model2[,2]),
            S.Thick = soil.thick_Model1, pClay = clay_Model1, # Same parameters as Section1
            pE = 1.0, bare = TRUE)$b # Dry season = no vegetation covering the soil (bare = T)

# Replace negative values in fW with 0
fW_Model2[fW_Model2 < 0] <- 0

# >>>ADD COMMENT<<
xi.frame_Model2=data.frame(years_Model2,rep(fT_Model2*fW_Model2,length.out=length(years_Model2)))

# Estimating IOM using Falloon method
FallIOM=0.049*SOC_Model1^(1.139)
```

```{r Model Model}
# Section1: Run the model
Model_Model1=RothCModel(t=years_Model1,C0=poolSize_Spinup, # Using pool size from spin-up
                  In=Cinputs_Model1, clay=clay_Model1, xi=xi.frame_Model1,
                  DR = DPM.RPM, FYM = FYM)
Ct_Model1=getC(Model_Model1) # Calculates SOC stocks for each pool

# Section2: Run the model
Model_Model2=RothCModel(t=years_Model2,C0=poolSize_Spinup, # Using pool size from spin-up
                  In=Cinputs_Model2, clay=clay_Model1, xi=xi.frame_Model2, # Only Cinputs change
                  DR = DPM.RPM, FYM = FYM)
Ct_Model2=getC(Model_Model2) # Calculates SOC stocks for each pool

# Section1: Final pool size 
poolSize_Model1=as.numeric(tail(Ct_Model1,1))
names(poolSize_Model1)<-c("DPM", "RPM", "BIO", "HUM", "IOM")
poolSize_Model1

# Section2: Final pool size 
poolSize_Model2=as.numeric(tail(Ct_Model2,1))
names(poolSize_Model2)<-c("DPM", "RPM", "BIO", "HUM", "IOM")
poolSize_Model2

# Sum of SOC stocks for both sections
rows_C_Model1 <- rowSums(Ct_Model1)
rows_C_Model2 <- rowSums(Ct_Model2)

# Sum of SOC stocks for both sections
total_C_Model1 <- sum(poolSize_Model1)
total_C_Model2 <- sum(poolSize_Model2)

# Create a data frame for the final pool sizes
final_pools_df <- data.frame(
    Pool = c("DPM", "RPM", "BIO", "HUM", "IOM"),
    Rainy_season = poolSize_Model1,
    Dry_season = poolSize_Model2,
    Spin_up = poolSize_Spinup
)

# Combine total carbon stocks into a data frame with spin-up
total_C_df <- data.frame(
    Model = c("Rainy_season", "Dry_season", "Spin_up"),
    Total_C = c(sum(poolSize_Model1), sum(poolSize_Model2), sum(poolSize_Spinup)),
    Years = c(10, 10, 500)
    )

# Print results
print(total_C_df)

# Determine the maximum y-axis limit to fit both lines comfortably
max_C_Model1 <- max(total_C_Model1)
```

```{r Plot Model}
# Plot the final pool sizes
ggplot(final_pools_df, aes(x = Pool)) +
    geom_bar(aes(y = Rainy_season, fill = "Rainy_season"), stat = "identity", position = "dodge") +
    geom_bar(aes(y = Dry_season, fill = "Dry_season"), stat = "identity", position = "dodge") +
    labs(title = "Final Pool Sizes for Rainy_season and Dry_season",
         x = "Carbon Pool",
         y = "Carbon Stock (Mg/ha)",
         fill = "Model") +
    theme_minimal()

# Plot total carbon stocks
ggplot(total_C_df, aes(x = Model, y = Total_C, fill = Model)) +
    geom_bar(stat = "identity") +
    labs(title = "Total Carbon Stocks for Rainy_season and Dry_season",
         x = "Model",
         y = "Total Carbon Stock (Mg/ha)") +
    theme_minimal()

# Plot individual pools and total SOC for the spin-up period
matplot(years_, Ct_Spinup, type = "l", lty = 1, col = 1:5,
        xlab = "Time (years)", ylab = "C stocks (Mg/ha)",
        main = "SOC Pools and Total (Spin-Up)", ylim = c(0, max_C_Model1 * 1.1))

lines(years_Spinup, total_C_spinup, col = "black", lty = 2)
legend("topright", c("Total SOC", "DPM", "RPM", "BIO", "HUM", "IOM"),
       lty = c(1, 1, 1, 1, 1, 1), col = c(1:5, "black"), bty = "n")
```

#-- Dump it --------------------------------------------------------------------

```{r Dumpster}
#!# values from ERA5 tiff are all negative, calculation error?
## seems like evaporation = negative values and condensation = positive values according to ERA5 information
## all values are in m > *-1000 to convert to positive values in mm

# Original values from ERA5
# Evp=data.frame(Month=1:12, Evp=c(-0.3, -0.3, -0.3, -0.3, -0.26, -0.24, -0.19, -0.15, -0.15, -0.22, -0.26, -0.32))

# ## plot results
# matplot(years_Spinup, Ct_Spinup, type="l", lty=1, col=1:5,
#         xlab="Time (years)", ylab="C stocks (Mg/ha)")
# legend("topleft", c("DPM", "RPM", "BIO", "HUM", "IOM"),
#        lty=1, col=1:5, bty="n")
```

Not working anymore
```{r Plot Spin-up}
# # Plot the final pool sizes
# ggplot(final_pools_df, aes(x = Pool)) +
#     geom_bar(aes(y = Rainy_season, fill = "Rainy_season"), stat = "identity", position = "dodge") +
#     geom_bar(aes(y = Dry_season, fill = "Dry_season"), stat = "identity", position = "dodge") +
#     labs(title = "Final Pool Sizes for Rainy_season and Dry_season",
#          x = "Carbon Pool",
#          y = "Carbon Stock (Mg/ha)",
#          fill = "Model") +
#     theme_minimal()

# Plot total carbon stocks
ggplot(total_C_df, aes(x = Model, y = Total_C, fill = Model)) +
    geom_bar(stat = "identity") +
    labs(title = "Total Carbon Stocks for Rainy_season and Dry_season",
         x = "Model",
         y = "Total Carbon Stock (Mg/ha)") +
    theme_minimal()

# Plot individual pools and total SOC for the spin-up period
matplot(years_Spinup, Ct_Spinup, type = "l", lty = 1, col = 1:5,
        xlab = "Time (years)", ylab = "C stocks (Mg/ha)",
        main = "SOC Pools and Total (Spin-Up)")

lines(years_Spinup, total_C_spinup, col = "black", lty = 2) # Not working anymore...
legend("topright", c("Total SOC", "DPM", "RPM", "BIO", "HUM", "IOM"),
       lty = c(1, 1, 1, 1, 1, 1), col = c(1:5, "black"), bty = "n")
```

```{r Model Experiment}
## calculate effects of climate on decomposition
fT=fT.RothC(Temp[,2]) #Temperature effects per month
## somehow take care of vegetation period / rain season i.e. bare = FALSe OR TRUE changes during year
fW=fW.RothC(P=(Precip[,2]), E=(Evp[,2]), 
            S.Thick = soil.thick, pClay = clay, 
            pE = 1.0, bare = FALSE)$b #Moisture effects per month

# Replace negative values in fW with 0
fW[fW < 0] <- 0

xi.frame=data.frame(years,rep(fT*fW,length.out=length(years)))

# Falloon is the standard method, other options are Zimmermann et al. (2017) fractionation or pedotransfer functions
FallIOM=0.049*SOC^(1.139) #IOM using Falloon method

## run the model for BAU scenario
Model_BAU=RothCModel(t=years,C0=poolSize_Spinup,
                  In=Cinputs_BAU, clay=clay, xi=xi.frame,
                  DR = DPM.RPM, FYM = FYM) #Loads the model
Ct_BAU=getC(Model_BAU) #Calculates stocks for each pool per month

## run the model for IALM scenario
Model_SALM=RothCModel(t=years,C0=poolSize_Spinup, ## In refers to litter, change FYM if you want to add manure
                  In=Cinputs_SALM, clay=clay, xi=xi.frame) #Loads the model 
Ct_SALM=getC(Model_SALM) #Calculates stocks for each pool per month

# Sum the carbon stocks for each scenario
total_C_BAU <- rowSums(Ct_BAU)
total_C_SALM <- rowSums(Ct_SALM)

# Calculate the difference in total SOC at the end of the time period
difference_total_C <- tail(total_C_SALM, 1) - tail(total_C_BAU, 1)

# Display the result
difference_total_C

# Determine the maximum y-axis limit to fit both lines comfortably
max_C_plot <- max(total_C_SALM, total_C_BAU)
```

```{r Plot Experiment}
# Plot the total SOC for SALM and BAU scenarios (order changed to SALM first for better line visibility)
plot(years, total_C_SALM, type = "l", col = "red", lty = 2, 
     xlab = "Time (years)", ylab = "Total C stocks (Mg/ha)",
     main = "SOC Stocks Comparison: BAU vs SALM", ylim = c(0, max_C_plot * 1.1))

# Add the BAU line to the plot
lines(years, total_C_BAU, col = "blue", lty = 1)

# Add a legend
legend("topright", legend = c("SALM", "BAU"), col = c("red", "blue"), lty = c(2, 1), bty = "n")

# Display the difference in total SOC at the end on the plot
text(x = max(years) * 0.7, y = 20,
     labels = paste("Delta after 20 years:", round(difference_total_C, 2), "Mg/ha"), pos = 1, col = "black")
```
